+++ source/core/parameters.cc
@@ -514,15 +514,21 @@ namespace Parameters
   }
 
   void
-  SurfaceTensionParameters::parse_parameters(ParameterHandler &prm)
+  SurfaceTensionParameters::parse_parameters(
+    ParameterHandler                &prm,
+    const Parameters::Dimensionality dimensions)
   {
     surface_tension_coefficient = prm.get_double("surface tension coefficient");
-    T_0                         = prm.get_double("reference state temperature");
+    surface_tension_coefficient *= dimensions.surface_tension_scaling;
+    T_0 = prm.get_double("reference state temperature");
+    T_0 *= 1. / dimensions.temperature;
     surface_tension_gradient =
       prm.get_double("temperature-driven surface tension gradient");
-    T_solidus  = prm.get_double("solidus temperature");
+    surface_tension_gradient *= dimensions.surface_tension_gradient_scaling;
+    T_solidus = prm.get_double("solidus temperature");
+    T_solidus *= 1. / dimensions.temperature;
     T_liquidus = prm.get_double("liquidus temperature");
-
+    T_liquidus *= 1. / dimensions.temperature;
     Assert(T_liquidus > T_solidus,
            PhaseChangeIntervalError(T_liquidus, T_solidus));
   }
@@ -539,10 +545,14 @@ namespace Parameters
   }
 
   void
-  MobilityCahnHilliardParameters::parse_parameters(ParameterHandler &prm)
+  MobilityCahnHilliardParameters::parse_parameters(
+    ParameterHandler                &prm,
+    const Parameters::Dimensionality dimensions)
   {
     mobility_cahn_hilliard_constant =
       prm.get_double("cahn hilliard mobility constant");
+    mobility_cahn_hilliard_constant *=
+      dimensions.cahn_hilliard_mobility_scaling;
   }
 
   template <int dim>
@@ -962,7 +972,7 @@ namespace Parameters
            ++i_material_interaction)
         {
           material_interactions[i_material_interaction].parse_parameters(
-            prm, i_material_interaction);
+            prm, i_material_interaction, dimensions);
           if (material_interactions[i_material_interaction]
                 .material_interaction_type ==
               MaterialInteractions::MaterialInteractionsType::fluid_fluid)
@@ -1304,7 +1314,10 @@ namespace Parameters
   }
 
   void
-  MaterialInteractions::parse_parameters(ParameterHandler &prm, unsigned int id)
+  MaterialInteractions::parse_parameters(
+    ParameterHandler                &prm,
+    unsigned int                     id,
+    const Parameters::Dimensionality dimensions)
   {
     prm.enter_subsection("material interaction " +
                          Utilities::int_to_string(id, 1));
@@ -1339,39 +1352,36 @@ namespace Parameters
             if (op == "constant")
               {
                 surface_tension_model = SurfaceTensionModel::constant;
-                surface_tension_parameters.parse_parameters(prm);
               }
             else if (op == "linear")
               {
                 surface_tension_model = SurfaceTensionModel::linear;
-                surface_tension_parameters.parse_parameters(prm);
               }
             else if (op == "phase change")
               {
                 surface_tension_model = SurfaceTensionModel::phase_change;
-                surface_tension_parameters.parse_parameters(prm);
               }
             else
               throw(std::runtime_error(
                 "Invalid surface tension model. The choices are <constant|linear|phase change>."));
-
+            surface_tension_parameters.parse_parameters(prm, dimensions);
             // Cahn-Hilliard mobility
             op = prm.get("cahn hilliard mobility model");
             if (op == "constant")
               {
                 mobility_cahn_hilliard_model =
                   MobilityCahnHilliardModel::constant;
-                mobility_cahn_hilliard_parameters.parse_parameters(prm);
               }
             else if (op == "quartic")
               {
                 mobility_cahn_hilliard_model =
                   MobilityCahnHilliardModel::quartic;
-                mobility_cahn_hilliard_parameters.parse_parameters(prm);
               }
             else
               throw(std::runtime_error(
                 "Invalid mobility model. The choices are <constant|quartic>."));
+
+            mobility_cahn_hilliard_parameters.parse_parameters(prm, dimensions);
           }
           prm.leave_subsection();
         }
@@ -1390,17 +1400,17 @@ namespace Parameters
           if (op == "constant")
             {
               surface_tension_model = SurfaceTensionModel::constant;
-              surface_tension_parameters.parse_parameters(prm);
+              surface_tension_parameters.parse_parameters(prm, dimensions);
             }
           else if (op == "linear")
             {
               surface_tension_model = SurfaceTensionModel::linear;
-              surface_tension_parameters.parse_parameters(prm);
+              surface_tension_parameters.parse_parameters(prm, dimensions);
             }
           else if (op == "phase change")
             {
               surface_tension_model = SurfaceTensionModel::phase_change;
-              surface_tension_parameters.parse_parameters(prm);
+              surface_tension_parameters.parse_parameters(prm, dimensions);
             }
           else
             throw(std::runtime_error(